feat: ReceiveModal 컴포넌트 분리
